Add s3_compat storage option for all S3 compatible storage targets
