+++ medusa/storage/s3_compat_storage/awscli.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright 2020 DataStax Inc.
+# Copyright 2019 The Last Pickle
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,15 +22,14 @@ import sys
 
 from retrying import retry
 
-from medusa.libcloud.storage.drivers.ibm import IBM_CLOUD_HOSTS_BY_REGION
-
 
 class AwsCli(object):
     def __init__(self, storage):
         self._config = storage.config
         self.storage = storage
-        self.endpoint_url = storage.config.host if storage.config.host is not None \
-            else IBM_CLOUD_HOSTS_BY_REGION[storage.config.region]
+        # Needs more parsing here - could allow IBM but doesn't need to
+        # self.endpoint_url = storage.config.host if storage.config.host is not None \
+        #     else IBM_CLOUD_HOSTS_BY_REGION[storage.config.region]
 
     @property
     def bucket_name(self):
@@ -77,8 +76,13 @@ class AwsCli(object):
         awscli_output = "/tmp/awscli_{0}.output".format(job_id)
         objects = []
         for src in srcs:
-            cmd = [self._aws_cli_path, "--endpoint-url", "https://{}".format(self.endpoint_url),
-                   "--region", self._config.region, "s3", "cp", str(src), "s3://{}/{}".format(bucket_name, dest)]
+            # TODO self.endpoint_url should be optional, used if available - otherwise use the s3 format
+            #      even s3 needs region support of course
+            # cmd = [self._aws_cli_path, "--endpoint-url", "https://{}".format(self.endpoint_url),
+            #        "--region", self._config.region, "s3", "cp", str(src), "s3://{}/{}".format(bucket_name, dest)]
+
+
+            cmd = [self._aws_cli_path, "s3", "cp", str(src), "s3://{}/{}".format(bucket_name, dest)]
             objects.append(self.upload_file(cmd, dest, awscli_output))
 
         return objects
@@ -87,8 +91,12 @@ class AwsCli(object):
         job_id = str(uuid.uuid4())
         awscli_output = "/tmp/awscli_{0}.output".format(job_id)
         objects = []
-        cmd = [self._aws_cli_path, "--endpoint-url", "https://{}".format(self.endpoint_url), "--region",
-               self._config.region, "s3", "cp", "s3://{}/{}".format(bucket_name, src), dest]
+        # TODO self.endpoint_url should be optional, used if available - otherwise use the s3 format
+        #      even s3 needs region support of course
+        # cmd = [self._aws_cli_path, "--endpoint-url", "https://{}".format(self.endpoint_url), "--region",
+        #        self._config.region, "s3", "cp", "s3://{}/{}".format(bucket_name, src), dest]
+
+        cmd = [self._aws_cli_path, "s3", "cp", "s3://{}/{}".format(bucket_name, src), dest]
         self.download_file(cmd, dest, awscli_output)
 
         return objects
